Forking joho/godotenv to improve performance
This commits improves the parser to use significantly fewer allocations
by utilizing a parser instead of regular expressions. The error
handling has been improved as well by using custom errors which includes
the line number and character at which the errors occur in addition to
visual cues of where the error is.

In addition many golangci lint issues have been resolved, something that
will be introduced in the next commit.

The `godotenv` binary has an improved help section, this is in
addition to the adding version information, and better handling of
multiple `.env` files.

The README has been updated for the fork, including a disclaimer that
this is as such.

To fix spelling issues, `LICENCE` has been renamed to `LICENSE`.

Lastly, `renovate.json` as been removed as its use is unknown.