godotenv binary now execs on non-Windows
Using exec is better in this case as it makes the process tree simpler,
and allows the child's signals to be handled by the caller. In addition
the `Exec` function has been removed as it's not really great as a
library function. Users should manually call the `Load` function and
run `exec.Command` or `syscall.Exec` appropriately.

In the future we probably want to handle signals for Windows anyway.

This is based on https://github.com/joho/godotenv/pull/77 and also
addresses the following PRs in upstream:

- https://github.com/joho/godotenv/pull/119